The verification process is a process that does not need an extra service.
